[kernel] E: Create enum for Multiboot tag type
