+++ src/hal/platform/pc/mboot/basic_mem_info.rs
@@ -5,6 +5,7 @@
 // Imports
 //==================================================================================================
 
+use crate::hal::platform::mboot::MbootTagType;
 use ::sys::error::{
     Error,
     ErrorCode,
@@ -22,7 +23,7 @@ use ::sys::error::{
 #[repr(C, align(8))]
 pub struct MbootBasicMeminfoTag {
     /// Type.
-    typ: u32,
+    typ: MbootTagType,
     /// Size.
     size: u32,
 }
